cleanups and file renames for consistency
