--- src/routes/terms-of-service/OnboardingTOC.svelte
@@ -1,5 +1,0 @@
-<script>
-  import Copy from "./Copy.svelte";
-</script>
-
-<Copy />
