+++ stories/flow/OnboardingFlowView.svelte
@@ -2,9 +2,9 @@
   import { onMount } from "svelte";
   import Layout from "../../src/components/Layout.svelte";
   import Main from "../../src/components/Main.svelte";
-  import Welcome from "../../src/routes/welcome/Welcome.svelte";
-  import Demographics from "../../src/routes/demographics/Body.svelte";
-  import TermsOfService from "../../src/routes/terms-of-service/OnboardingTOC.svelte";
+  import Welcome from "../../src/routes/welcome/Content.svelte";
+  import Demographics from "../../src/routes/demographics/Content.svelte";
+  import TermsOfService from "../../src/routes/terms-of-service/Content.svelte";
   import OnboardingCTAContainer from "../../src/components/OnboardingCTAContainer.svelte";
 
   import DemographicsCallToAction from "../../src/routes/demographics/CallToAction.svelte";
