+++ stories/flow/TermsView.svelte
@@ -1,9 +1,8 @@
 <script>
   import { onMount } from "svelte";
-  import { fly } from "svelte/transition";
   import Layout from "../../src/components/Layout.svelte";
   import Main from "../../src/components/Main.svelte";
-  import Terms from "../../src/routes/terms-of-service/OnboardingTOC.svelte";
+  import TermsOfService from "../../src/routes/terms-of-service/Content.svelte";
   import OnboardingCTAContainer from "../../src/components/OnboardingCTAContainer.svelte";
   import CallToAction from "../../src/routes/terms-of-service/CallToAction.svelte";
 
@@ -14,10 +13,10 @@
 </script>
 
 {#if mounted}
-  <div in:fly={{ duration: 800, y: 5 }}>
+  <div>
     <Layout>
       <Main padForOnboarding>
-        <Terms />
+        <TermsOfService />
       </Main>
       <OnboardingCTAContainer step={2}>
         <CallToAction
